requirements.txt
# Core data processing and analysis
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Visualization
plotly>=5.5.0
umap-learn>=0.5.2
pillow>=9.0.0

# Data acquisition
gdown>=4.4.0

# Excel file handling
openpyxl>=3.0.9
xlrd>=2.0.1